+++ src/game/Object/PlayerPvP.cpp
@@ -1,0 +1,181 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+#include "Player.h"
+#include "Language.h"
+#include "Database/DatabaseEnv.h"
+#include "Log.h"
+#include "Opcodes.h"
+#include "SpellMgr.h"
+#include "World.h"
+#include "WorldPacket.h"
+#include "WorldSession.h"
+#include "UpdateMask.h"
+#include "SkillDiscovery.h"
+#include "QuestDef.h"
+#include "GossipDef.h"
+#include "UpdateData.h"
+#include "Channel.h"
+#include "ChannelMgr.h"
+#include "MapManager.h"
+#include "MapPersistentStateMgr.h"
+#include "InstanceData.h"
+#include "GridNotifiers.h"
+#include "GridNotifiersImpl.h"
+#include "CellImpl.h"
+#include "ObjectMgr.h"
+#include "ObjectAccessor.h"
+#include "CreatureAI.h"
+#include "Formulas.h"
+#include "Group.h"
+#include "Guild.h"
+#include "GuildMgr.h"
+#include "Pet.h"
+#include "Util.h"
+#include "Transports.h"
+#include "Weather.h"
+#include "BattleGround/BattleGround.h"
+#include "BattleGround/BattleGroundMgr.h"
+#include "BattleGround/BattleGroundAV.h"
+#include "OutdoorPvP/OutdoorPvP.h"
+#include "ArenaTeam.h"
+#include "Chat.h"
+#include "revision_data.h"
+#include "Database/DatabaseImpl.h"
+#include "Spell.h"
+#include "ScriptMgr.h"
+#include "SocialMgr.h"
+#include "AchievementMgr.h"
+#include "Mail.h"
+#include "SpellAuras.h"
+#include "DBCStores.h"
+#include "DB2Stores.h"
+#include "SQLStorages.h"
+#include "Vehicle.h"
+#include "Calendar.h"
+#include "DisableMgr.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+
+#include <cmath>
+/// checks the 15 afk reports per 5 minutes limit
+void Player::UpdateAfkReport(time_t currTime)
+{
+    if (m_bgData.bgAfkReportedTimer <= currTime)
+    {
+        m_bgData.bgAfkReportedCount = 0;
+        m_bgData.bgAfkReportedTimer = currTime + 5 * MINUTE;
+    }
+}
+
+void Player::UpdateContestedPvP(uint32 diff)
+{
+    if (!m_contestedPvPTimer || IsInCombat())
+    {
+        return;
+    }
+    if (m_contestedPvPTimer <= diff)
+    {
+        ResetContestedPvP();
+    }
+    else
+    {
+        m_contestedPvPTimer -= diff;
+    }
+}
+
+/**
+ * @brief Updates and clears the player's PvP flag when the timeout expires.
+ *
+ * @param currTime The current server time.
+ */
+void Player::UpdatePvPFlag(time_t currTime)
+{
+    if (!IsPvP())
+    {
+        return;
+    }
+    if (pvpInfo.endTimer == 0 || currTime < (pvpInfo.endTimer + 300))
+    {
+        return;
+    }
+
+    UpdatePvP(false);
+}
+
+/**
+ * @brief Starts an active duel once the duel countdown completes.
+ *
+ * @param currTime The current server time.
+ */
+void Player::UpdateDuelFlag(time_t currTime)
+{
+    if (!duel || duel->startTimer == 0 || currTime < duel->startTimer + 3)
+    {
+        return;
+    }
+
+    // Used by Eluna
+#ifdef ENABLE_ELUNA
+    if (Eluna* e = GetEluna())
+    {
+        e->OnDuelStart(this, duel->opponent);
+    }
+#endif /* ENABLE_ELUNA */
+
+    SetUInt32Value(PLAYER_DUEL_TEAM, 1);
+    duel->opponent->SetUInt32Value(PLAYER_DUEL_TEAM, 2);
+
+    duel->startTimer = 0;
+    duel->startTime  = currTime;
+    duel->opponent->duel->startTimer = 0;
+    duel->opponent->duel->startTime  = currTime;
+}
+
+/**
+ * @brief Sets or clears the player's PvP state with timeout-aware handling.
+ *
+ * @param state True to enable PvP; false to disable it.
+ * @param ovrride True to bypass the delayed PvP timeout behavior.
+ */
+void Player::UpdatePvP(bool state, bool ovrride)
+{
+    if (!state || ovrride)
+    {
+        SetPvP(state);
+        pvpInfo.endTimer = 0;
+    }
+    else
+    {
+        if (pvpInfo.endTimer != 0)
+        {
+            pvpInfo.endTimer = time(NULL);
+        }
+        else
+        {
+            SetPvP(state);
+        }
+    }
+}
